intes

+++ src/data/CartaSup.java
@@ -3,7 +3,7 @@ package src.data;
 public class CartaSup extends CartaP {
     private String efeito;
 
-    public CartaSup(String nome, int tipo, double poder, String efeito, String descricao) {
+    public CartaSup(String nome, int tipo, int poder, String efeito, String descricao) {
         super(nome, "SUPORTE", poder, efeito, descricao); 
         this.efeito = efeito;
     }
