+++ src/pannel/GerenciadorCartas.java
@@ -48,7 +48,7 @@ public class GerenciadorCartas {
             while ((linha = br.readLine()) != null) {
                 String[] c = linha.split(",");
 
-                CartaSup carta = new CartaSup(c[0], Double.parseDouble(c[2]), Integer.parseInt(c[3]), c[4], c[5]);
+                CartaSup carta = new CartaSup(c[0], Integer.parseInt(c[2]), Integer.parseInt(c[3]), c[4], c[5]);
                 cartas.put(carta.getNome(), carta);
             }
         } catch (Exception e) {}
